+++ native/react/src/components/molecules/SignInForm/index.tsx
@@ -28,7 +28,7 @@ export default function SignInForm(props: Props) {
   return (
     <>
       <TextField placeholder="Email" value={email.value} onChangeText={email.onChangeText} />
-      <TextField placeholder="Password" value={password.value} onChangeText={password.onChangeText} />
+      <TextField placeholder="Password" value={password.value} onChangeText={password.onChangeText} secureTextEntry={true} />
       <Button
         onPress={handleSignIn}
         title="ログイン"
