+++ native/react/src/components/organisms/index.ts
@@ -1,2 +1,3 @@
 export { default as Profile } from "./Profile";
+export { default as SignInCard } from "./SignInCard";
 export { default as UserInfoList } from "./UserInfoList";
